Changed the code-style in authlogic.
Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.

Changed the code-style in authlogic.